+++ src/fs-promise-proxy.js
